Generics, the best of polymorphism.
With the wildcard( ? ) the sintax that representive the polymorphism is more simple .

Instead of having several services and tests for different classes (domains), how about just changing the type (object) and having the same functionalities?

How to do it´s?
In my case, I copy one class of service of many that have in packet of service, and that they are equal in terms of the result. Right away, the construct of new class have one List of type Generic(the object), and in construct have the passage of list to member of class, that representive the object that we working.

So, in criation of object of reperesentive a service, I defination the type. For example; RentalService<Carro> rentalService = new RentalService(carrosDisponiveis);