Is it possible to instantiate an interface in Java?
In this example, I needed to use anonymous classes to reference the type of the interface, so, indirectly, I instantiated the interface.

Moreover, is it possible to override the behavior of a method in a class with anonymous classes only temporarily?

Yes, and in addition to that, there is a considerable reduction in code for achieving the same goal, compared to overriding using the class object that contains the method.

About other classes, I observed that:

1️⃣ Inner Classes:
These are excellent for functionalities of graphical components. For example, when class B (and only it) needs to access the attributes of class A.
To access the components of an inner class, you need an object of the outer class.

2️⃣ Local Classes:
Basically, these are classes declared inside a method, or even inside a static initialization block.
The method that contains the local class in its body needs to instantiate the local class within itself. This is because, once the method ends, there will be no way to access the local class anymore.

Important Notes:

The local class can be either abstract or final (access modifiers).
The attributes and parameters of the method must be final or effectively final. This is because the local class might remain in memory, but the method that contains it will no longer exist.
3️⃣ What’s the difference between an inner class and a static inner class?

The main difference is that by making an inner class static, it behaves as a top-level class.
To access members of the outer class, you need to either explicitly mention it or instantiate it within the static class.

The primary purpose of a static inner class is packaging related functionalities. Moreover, it is possible to directly access the static inner class in a main method.