"Have you ever come across methods that repeat almost the same behavior but with slight differences? How about simplifying them and making them more dynamic with parameterization?"
In this example, I’m going to parameterize a car filter lol, it might sound crazy, but the idea is simple. Imagine a hypothetical scenario where you receive a database (DB) list of cars with different colors and years, and you need to filter them by these two characteristics.

At first, I create methods for different colors and manufacturing years. Then, I add the color (`String`) and year (`int`) as the second parameter, reducing the number of methods.
But it’s clear that only the logic inside the `if()` is being changed! What if we passed this condition as a parameter?

To achieve this, I create an interface (polymorphism is a great friend) with a method inside (`test`) without behavior, and I define the behavior using an anonymous class (a subclass of `CarPredicate`).

**Note:** In this case, we’re still using the lambda syntax (available since Java 8).

**Summary:** We define just one logic for the business rule, which will change with the call (parameter)!

**Would it be possible to create a method that filters anything?**

Yes, but for this, we use the `Predicate` interface (designed for generalization) and modify the method's syntax, replacing the `Car` type with a generic type `<T>`, defined at the time of the call.