Have you ever heard of lambda in Java or even in Mathematics? I said lambda, not "licking" (lambida) 😂.
Jokes aside, lambdas came to make code easier to understand and reduce the number of lines, because before Java 8, we basically had to use anonymous classes to define rules for standardizing behavior, for example.

There are a few ways to work with lambdas:

Note: Lambdas work with functional interfaces (interfaces with only one abstract method), as these interfaces define how the lambda should be used. Also, lambdas are anonymous, and since they are not tied to a class, they act as "functions" and are concise.

Interface + explanatory text about the lambda’s behavior.

Predicate: In the case of the Predicate interface, when implementing the logic, you must consider a return type (boolean), and its syntax consists of: parameter + return type -> expression.

Consumer: There are cases where the lambda logic has no return, so there is an appropriate interface for that, the Consumer interface, with its method void accept(T t);.

Function: Think of a scenario where you pass a list of Strings and want to get the length of each string in the list. For this, in the parameters, you need to pass:

1st: A List<String>.
2nd: The logic followed by the variable, e.g.:
List<Integer> lengths = map(strings, s -> s.length()); (*Here, strings refers to the List<String>).
Note: The T and R types can be the same. Moreover, I found it clever that the IntelliJ IDE, when implementing the logic via lambda (and if I hadn’t yet defined the variable type that would receive the return of the map method), shows the option to use Alt + Enter to introduce a local variable. The IDE automatically infers the type of the variable based on the logic being passed, whether it’s an integer, String, or any other type.